chore: extract walk_dir_but_exclude_some helper function
