+++ src/layout/mod.rs
@@ -1,6 +1,6 @@
 //! 启发式了解项目的 Rust packages 组织结构。
 
-use crate::Result;
+use crate::{utils::walk_dir_but_exclude_some, Result};
 use cargo_metadata::{
     camino::{Utf8Path, Utf8PathBuf},
     Metadata, MetadataCommand,
@@ -16,16 +16,7 @@ use cargo_check_verbose::PackageInfo;
 
 /// 寻找仓库内所有 Cargo.toml 所在的路径
 fn find_all_cargo_toml_paths(repo_root: &str, dirs_excluded: &[&str]) -> Vec<Utf8PathBuf> {
-    let mut cargo_tomls: Vec<Utf8PathBuf> = walkdir::WalkDir::new(repo_root)
-        .max_depth(10) // 目录递归上限
-        .into_iter()
-        .filter_entry(|entry| {
-            // 别进入这些文件夹（适用于子目录递归）
-            const NO_JUMP_IN: &[&str] = &[".git", "target"];
-            let filename = entry.file_name();
-            let excluded = &mut NO_JUMP_IN.iter().chain(dirs_excluded);
-            !excluded.any(|&dir| dir == filename)
-        })
+    let mut cargo_tomls: Vec<Utf8PathBuf> = walk_dir_but_exclude_some(repo_root, 10, dirs_excluded)
         .filter_map(|entry| {
             // 只搜索 Cargo.toml 文件
             let entry = entry.ok()?;
